requirements.txt
python-dotenv
selenium

